requirements.txt
discord.py
aioschedule
